feat(tooling): implement v0.16.0 Developer Tooling APIs
Phase 18 implementation providing comprehensive developer tooling:

- Project Scaffolding API (project.zig)
  - Create project layouts for 7 platforms (iOS, Android, Web, macOS, Windows, Linux, Embedded)
  - Template-based initialization with ProjectConfig
  - Validation and project metadata management
  - Future-based async operations

- Build Orchestration API (build.zig)
  - Multi-target build execution with BuildMode (debug/release/distribution)
  - Progress and log streaming with callbacks
  - Parallel build support
  - Build state management (pending/compiling/linking/packaging/completed/failed/cancelled)

- Build Artifact Query API (artifacts.zig)
  - Artifact path retrieval and metadata access
  - SHA-256 hash verification
  - Export functionality with compression options
  - Platform-specific extensions (ipa/apk/app/wasm/exe)

- Target Capability Matrix API (targets.zig)
  - 30+ features mapped across platforms
  - Support levels (full/partial/none/native)
  - Platform-specific input requirements (bundle_id, package_name)
  - Feature compatibility validation

- Template Catalog API (templates.zig)
  - 8 built-in templates (app, game-2d, game-3d, library, widget, embedded, server, plugin)
  - Custom template registration
  - Template versioning and metadata

- File Watcher API (watcher.zig)
  - Real-time file system monitoring
  - Configurable filters (patterns, hidden files, build dirs)
  - Debounced change events
  - Pause/resume functionality

- Component Tree Export API (ui.zig)
  - Extract component hierarchy
  - JSON/YAML/XML export formats
  - ComponentType enumeration (view, button, text, image, etc.)
  - Property and binding information

- Live Preview Bridge API (preview.zig)
  - Launch preview sessions per platform
  - Hot reload integration
  - Debug overlay support (bounds, margins, padding)
  - Multi-device simulation

Also includes:
- Integration into main.zig
- ToolingManager for unified access
- Comprehensive unit tests
- Memory leak fixes in test code
- ROADMAP.md updated to mark Phase 18 complete

+++ core/src/integration/keyvalue.zig
@@ -449,7 +449,7 @@ pub const BatchOperation = struct {
             switch (op) {
                 .put_bool => |o| self.store.putBool(o.key, o.value),
                 .put_int => |o| self.store.putInt(o.key, o.value),
-                .put_float => |o| self.store.putFloat(o.key, o.value),
+                .put_float => |o| self.store.putDouble(o.key, o.value),
                 .put_string => |o| self.store.putString(o.key, o.value),
                 .remove => |key| _ = self.store.remove(key),
             }
